requirements.txt
opencv-python==4.12.0.88
pytesseract==0.3.13
numpy==2.2.6
scipy==1.16.1
numba==0.61.2
torch==2.8.0+cpu
torchvision==0.23.0+cpu
pygame==2.6.1
customtkinter==5.2.2
pillow==11.3.0
pyyaml==6.0.2
matplotlib==3.10.6
seaborn==0.13.2
mss==10.1.0
psutil==7.0.0